change fields to floats

+++ db/migrate/20150825202238_create_people.rb
@@ -1,8 +1,8 @@
 class CreatePeople < ActiveRecord::Migration
   def change
     create_table :people do |t|
-      t.integer :height
-      t.integer :weight
+      t.float :height
+      t.float :weight
       t.string :gender
 
       t.timestamps null: false
